Tidy closed sockets from listeners event loop
This is related to https://github.com/net-ssh/net-ssh-gateway/issues/4. I traced the problem to Net::SSH::Compat.io_select which hangs indefinitely when it encounters a closed socket.. which occurs in forward.rb:L94 when the on_open_failed closes the socket. Not sure if this is best place to fix but checking for sockets that are closed and getting rid of them before calling io_select, it fixed the problem i had.